Fasten python installation checks when already installed

+++ tools/python/setup.sh
@@ -61,14 +61,17 @@ function setup_python() {
     }
   fi
 
-  local python_modules=$(git --no-pager config -f "$HOME/.common_env.ini" --get-all python.modules 2>/dev/null)
+  local python_modules
+  python_modules=$(git --no-pager config -f "$HOME/.common_env.ini" --get-all python.modules 2>/dev/null)
   for py in "$python_bin" "$APPS_ROOT/home/.venv/3/bin/python" "$APPS_ROOT/home/.venv/$python_version/Scripts/python.exe"; do
     if [[ -f "$py" ]]; then
       [[ "$py" = "/usr/bin/python" ]] && continue
+      [[ "$(readlink "$py" | sed -re 's/^(.*)\.exe$/\1/')" = "/usr/bin/python" && -f "$(dirname "$py")/autopep8" ]] && continue
       echoColor 36 "Checking $py..."
       # ! "$py" -m autopep8 --version &>/dev/null && "$py" -m pip install --upgrade wheel pip pylint autopep8
       "$py" -m pip install --upgrade wheel pip setuptools
-      [[ -n "$python_modules" ]] && "$py" -m pip install --upgrade $(echo "$python_modules")
+      # shellcheck disable=SC2086
+      [[ -n "$python_modules" ]] && "$py" -m pip install --upgrade $python_modules
 
       # "$py" -m pip config set global.index-url "https://pypi.python.org/simple/"
       # "$py" -m pip config set global.find-links "https://pypi.python.org/simple/ https://pypi.org/simple/"
@@ -81,5 +84,4 @@ function setup_python() {
       # "$py" -m pip install --upgrade autopep8 #--user
     fi
   done
-  set +x
 }
